Remove GetOID and GetRawBytes from IPublicKey.
There is no sensible interpretation of public keys in general the IC spec which operates on the "raw bytes", i.e. the keytype-defendant representation. In order to work with the raw bytes, you must know the concrete key type, so each concrete key type class can provide these methods if it so desires.